+++ console2/src/components/organisms/index.ts
@@ -35,6 +35,7 @@ export { default as OrganizationActivity } from './OrganizationActivity';
 export { default as OrganizationList } from './OrganizationList';
 export { default as OrganizationListActivity } from './OrganizationListActivity';
 export { default as ProcessActivity } from './ProcessActivity';
+export { default as ProcessChildrenActivity } from './ProcessChildrenActivity';
 export { default as ProcessFormActivity } from './ProcessFormActivity';
 export { default as ProcessHistoryActivity } from './ProcessHistoryActivity';
 export { default as ProcessListActivity } from './ProcessListActivity';