+++ console2/src/state/data/processes/types.ts
@@ -31,6 +31,7 @@ import { RequestState } from '../common';
 import { State as LogState } from './logs/types';
 import { State as PollState } from './poll/types';
 import { State as historyState } from './history/types';
+import { State as ChildrenState } from './children/types';
 
 export interface GetProcessRequest extends Action {
     instanceId: ConcordId;
@@ -90,4 +91,5 @@ export interface State {
     log: LogState;
     poll: PollState;
     history: historyState;
+    children: ChildrenState;
 }