+++ it/server/src/test/java/com/walmartlabs/concord/it/server/GithubEventResourceIT.java
@@ -54,7 +54,7 @@ public class GithubEventResourceIT extends AbstractServerIT {
     private static final String GITHUB_WEBHOOK_SECRET = "12345";
     private static final String HMAC_SHA1_ALGORITHM = "HmacSHA1";
 
-    @Test(timeout = 60000)
+    @Test(timeout = 260000)
     public void pushUnknownRepo() throws Exception {
         String processTag = "tag_" + randomString();
         Path tmpDir = createTempDir();
@@ -112,7 +112,7 @@ public class GithubEventResourceIT extends AbstractServerIT {
         ProcessApi processApi = new ProcessApi(getApiClient());
 
         while (!Thread.currentThread().isInterrupted()) {
-            List<ProcessEntry> process = processApi.list(projectId, null, null, Collections.singletonList(processTag), ProcessEntry.StatusEnum.FINISHED.getValue(), null, expectedCount + 1);
+            List<ProcessEntry> process = processApi.list(projectId, null, null, Collections.singletonList(processTag), ProcessEntry.StatusEnum.FINISHED.getValue(), null, null, expectedCount + 1);
             if (process.size() == expectedCount) {
                 return process;
             }